Add BirthDate property and new Masking project
Updated the `Person` class to include a `BirthDate` property of type `DateOnly` and modified the XML documentation for the `CreditCardNumber` property.

Added a new project named "Masking" to the solution and updated project references for "EF_Core_MaskingSample" and "EF_Core_MaskingSample1" with new GUIDs to reflect reorganization.

by Karen